# docs/technexus-solution-closure/02-设计阶段/openapi.yaml
openapi: 3.1.0
info:
  title: technexus-server API
  version: 1.0.1
  x-service-name: technexus-server
  description: >-
    TechNexus V1 user and administration contract. Demand is not an Order and
    this contract does not expose payment, purchase, entitlement, video upload,
    or instant-messaging capabilities.
servers:
  - url: /api/v1
    description: Same-origin reverse-proxy base path
...
security:
  - bearerAuth: []
paths:
...
  /demands:
    get:
      tags: [Demand]
      operationId: listDemands
      x-requirements: [FR-DMD-01, FR-DMD-07, FR-SCH-02]
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: sort
          in: query
          schema: {type: string, enum: [relevance, latest, budget, deadline]}
      responses:
        '200':
          description: Visibility-filtered demand page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandPage'}
        '400': {$ref: '#/components/responses/Problem'}
    post:
      tags: [Demand]
      operationId: createDemand
      x-requirements: [FR-DMD-01, FR-DMD-07]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DemandWrite'}
      responses:
        '201':
          description: Demand draft created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /demands/{demandId}:
    parameters:
      - $ref: '#/components/parameters/DemandId'
    get:
      tags: [Demand]
      operationId: getDemand
      x-requirements: [FR-DMD-01, FR-DMD-07]
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Demand filtered to caller visibility
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandResponse'}
        '404': {$ref: '#/components/responses/Problem'}
    patch:
      tags: [Demand]
      operationId: updateDemand
      x-requirements: [FR-DMD-01, FR-DMD-06]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DemandPatch'}
      responses:
        '200':
          description: Demand draft updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /demands/{demandId}/submit:
    post:
      tags: [Demand]
      operationId: submitDemandForReview
      x-requirements: [FR-DMD-02, FR-AUD-01]
      parameters:
        - $ref: '#/components/parameters/DemandId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpectedVersionCommand'}
      responses:
        '202':
          description: Review accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandResponse'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /demands/{demandId}/proposals:
    post:
      tags: [Demand]
      operationId: createProposal
      x-requirements: [FR-DMD-03, FR-DMD-07]
      parameters:
        - $ref: '#/components/parameters/DemandId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProposalWrite'}
      responses:
        '201':
          description: Proposal submitted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProposalResponse'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /demands/{demandId}/transitions:
    post:
      tags: [Demand]
      operationId: transitionDemand
      x-requirements: [FR-DMD-04, FR-DMD-06]
      parameters:
        - $ref: '#/components/parameters/DemandId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DemandTransition'}
      responses:
        '200':
          description: Transition applied
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemandResponse'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: tn_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
...
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 1024}
    PageSize:
      name: pageSize
      in: query
      schema: {type: integer, minimum: 1, maximum: 50, default: 20}
...
    DemandId:
      name: demandId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Instant:
      type: string
      format: date-time
...
    DemandState:
      type: string
      enum: [DRAFT, PENDING_REVIEW, APPROVED, REJECTED, PUBLISHED, NEGOTIATING, CONFIRMED, IN_PROGRESS, DELIVERED, COMPLETED, EXPIRED, CANCELLED, OFFLINE]
...
    Visibility:
      type: string
      enum: [PUBLIC, LOGIN_REQUIRED, PARTICIPANTS, OWNER_ONLY, ADMIN_ONLY]
    Money:
      type: object
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount: {type: string, pattern: '^\d{1,10}(\.\d{1,2})?$'}
        currency: {type: string, const: CNY}
...
    ExpectedVersionCommand:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: {type: integer, minimum: 0}
...
    Demand:
      type: object
      additionalProperties: false
      required: [id, state, title, description, visibility, version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        state: {$ref: '#/components/schemas/DemandState'}
        title: {type: string, maxLength: 160}
        description: {type: string, maxLength: 50000}
        technicalRequirements: {type: array, items: {type: string, maxLength: 500}, maxItems: 30}
        budgetMin: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        budgetMax: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        platformQuote: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        finalQuote: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        deadlineAt: {oneOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        visibility: {$ref: '#/components/schemas/Visibility'}
        version: {type: integer, minimum: 0}
        auditTaskId:
          oneOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}]
          description: Present only when a review submission has been accepted.
    DemandWrite:
      type: object
      additionalProperties: false
      required: [title, description, visibility]
      properties:
        title: {type: string, minLength: 1, maxLength: 160}
        description: {type: string, minLength: 1, maxLength: 50000}
        technicalRequirements: {type: array, items: {type: string, maxLength: 500}, maxItems: 30}
        budgetMin: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        budgetMax: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        deadlineAt: {oneOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        visibility: {$ref: '#/components/schemas/Visibility'}
        contact: {type: string, maxLength: 1000, writeOnly: true}
        fileIds: {type: array, maxItems: 20, uniqueItems: true, items: {$ref: '#/components/schemas/Uuid'}}
    DemandPatch:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        title: {type: string, minLength: 1, maxLength: 160}
        description: {type: string, minLength: 1, maxLength: 50000}
        technicalRequirements: {type: array, items: {type: string, maxLength: 500}, maxItems: 30}
        budgetMin: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        budgetMax: {oneOf: [{$ref: '#/components/schemas/Money'}, {type: 'null'}]}
        deadlineAt: {oneOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        visibility: {$ref: '#/components/schemas/Visibility'}
        contact: {type: [string, 'null'], maxLength: 1000, writeOnly: true}
        fileIds: {type: array, maxItems: 20, uniqueItems: true, items: {$ref: '#/components/schemas/Uuid'}}
        expectedVersion: {type: integer, minimum: 0}
    DemandResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: '#/components/schemas/Demand'}
    DemandPage:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Demand'}}
        nextCursor: {type: [string, 'null']}
    ProposalWrite:
      type: object
      additionalProperties: false
      required: [plan, techStack, estimatedDays, suggestedQuote]
      properties:
        plan: {type: string, minLength: 1, maxLength: 50000}
        techStack: {type: array, minItems: 1, maxItems: 30, items: {type: string, maxLength: 80}}
        estimatedDays: {type: integer, minimum: 1, maximum: 3650}
        suggestedQuote: {$ref: '#/components/schemas/Money'}
    ProposalResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, demandId, state, estimatedDays, suggestedQuote, version]
          properties:
            id: {$ref: '#/components/schemas/Uuid'}
            demandId: {$ref: '#/components/schemas/Uuid'}
            state: {type: string, enum: [SUBMITTED, WITHDRAWN, ACCEPTED, REJECTED]}
            estimatedDays: {type: integer}
            suggestedQuote: {$ref: '#/components/schemas/Money'}
            version: {type: integer, minimum: 0}
    DemandTransition:
      type: object
      additionalProperties: false
      required: [action, expectedVersion]
      properties:
        action:
          type: string
          enum: [START_NEGOTIATION, CONFIRM_PROPOSAL, START_WORK, DELIVER, COMPLETE, CANCEL, OFFLINE]
        proposalId: {oneOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}]}
        reason: {type: string, maxLength: 1000}
        expectedVersion: {type: integer, minimum: 0}
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        detail: {type: string}
        traceId: {type: string}
        violations:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: {type: string}
              code: {type: string}
              message: {type: string}